restructure signup form page file structure
